+++ ChangeLog.markdown
@@ -6,7 +6,9 @@
 * Support (partial) for ColReorder (for filter_type 'select' and 'text') https://github.com/vedmack/yadcf/issues/107
 * Skip empty strings in the column (avoid entering empty strings into the select / auto_complete filters) https://github.com/vedmack/yadcf/pull/112
 * Support complex headers (rowspan / colspan) https://github.com/vedmack/yadcf/issues/119
-
+* Support custom_func / multi_select_custom_func in exResetAllFilters and exFilterColumn functions https://github.com/vedmack/yadcf/issues/126
+* Added initMultipleColumns function to allows to add filter for multiple columns on a table https://github.com/vedmack/yadcf/issues/79
+* Bugs / Issues fix https://github.com/vedmack/yadcf/issues/136 / https://github.com/vedmack/yadcf/issues/131
 
 ## 0.8.6